Extract pcap metadata parsing into subfolder
