+++ 2-Computer-Systems/problems/5-Syn-Flood/src/pcap/metadata.d.ts
@@ -1,0 +1,2 @@
+export type Endianness = 'BE' | 'LE';
+export type SubSecondsType = 'micro' | 'nano';
