+++ 2-Computer-Systems/problems/5-Syn-Flood/src/start.ts
@@ -1,28 +1,6 @@
 import { open, readFile } from 'fs/promises';
-
-const LE_MICRO_MAGIC_STRING_UINT = 2712847316;
-const BE_MICRO_MAGIC_STRING_UINT = 3569595041;
-const LE_NANO_MAGIC_STRING_UINT = 2712812621;
-const BE_NANO_MAGIC_STRING_UINT = 1295823521;
-
-type Endianness = 'BE' | 'LE';
-type SubSecondsType = 'micro' | 'nano';
-
-const getPCapMetadata = (fileBuffer: Buffer): [Endianness, SubSecondsType] => {
-  const magicStringUInt = fileBuffer.readUInt32LE(0);
-  switch (magicStringUInt) {
-    case LE_MICRO_MAGIC_STRING_UINT:
-      return ['LE', 'micro'];
-    case BE_MICRO_MAGIC_STRING_UINT:
-      return ['BE', 'micro'];
-    case LE_NANO_MAGIC_STRING_UINT:
-      return ['LE', 'nano'];
-    case BE_NANO_MAGIC_STRING_UINT:
-      return ['BE', 'nano'];
-    default:
-      throw new Error(`PCap file has unrecognised magic string ${magicStringUInt.toString(16)}`);
-  }
-};
+import getPCapMetadata from './pcap/getPcapMetadata';
+import { Endianness, SubSecondsType } from './pcap/metadata';
 
 const LINK_HEADER_OFFSET = 22;
 const FIRST_PACKET_RECORD_OFFSET = LINK_HEADER_OFFSET + 2;
